format files & move folder
