+++ TrashMob/client-app/src/components/Map/GoogleMarkerWithInfoWindow/index.ts
@@ -1,0 +1,1 @@
+export * from './GoogleMarkerWithInfoWindow'
