+++ TrashMob/client-app/src/components/Map/index.ts
@@ -1,1 +1,3 @@
-export { MarkerWithInfoWindow } from './google/MarkerWithInfoWindow'
+export { GoogleMarkerWithInfoWindow as MarkerWithInfoWindow } from './GoogleMarkerWithInfoWindow'
+
+export { AzureSearchLocationInput as SearchLocationInput } from './AzureSearchLocationInput'
